srl bot overhaul, this commit will be broken
